requirements.txt
boto3==1.9.149
botocore==1.12.149
dj-database-url==0.5.0
Django==2.2
django-ckeditor==5.7.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-storages==1.7.1
docutils==0.14
gunicorn==19.9.0
jmespath==0.9.4
Pillow==6.0.0
psycopg2==2.8.2
python-dateutil==2.8.0
pytz==2018.9
s3transfer==0.2.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.3
whitenoise==4.1.2
